changes:
1)InjectDemoResources

getParamsUsingContext() method is added.
here @context annotation is used to get hold of the param and access it.
This is used when we might not know the name of params before accessing
it. Like we might want to fetch each header and to see what is in each
header

2) Class MessageFilterBeans{} added

this calss is used to put all the query params at one place when its
seesms messy at the method (getMessage() in this case) of resource class
(Messageresource in this case).

Now we can access the params that is inside MessageFilterBeans class by
using its instanc in the method of resources class.