Lesson 24.
This class was not pushed in earlier commit. so pusing it in this
commit. 
Changes is same as mentioned in earlier commit.


) Class MessageFilterBeans{} added

this calss is used to put all the query params at one place when its
seesms messy at the method (getMessage() in this case) of resource class
(Messageresource in this case).

Now we can access the params that is inside MessageFilterBeans class by
using its instanc in the method of resources class.